requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
faiss-cpu
filelock
fsspec
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.20.3
idna==3.10
ipykernel==6.29.5
ipython
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
llvmlite
MarkupSafe
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx
nltk==3.9.1
numba
numpy
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow
platformdirs
prompt_toolkit==3.0.51
protobuf
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic
pydantic_core
Pygments==2.19.1
pynndescent==0.5.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn
scipy
sentence-transformers==2.3.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tenacity==8.2.3
threadpoolctl
tiktoken==0.5.1
tokenizers==0.15.2
torch
torchvision
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.36.2
triton
typing_extensions==4.13.2
umap-learn==0.5.5
urllib3==1.26.6
wcwidth==0.2.13